# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Social Seller Toolkit API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: niche
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /wishlist:
    get:
      operationId: listWishlist
      tags: [wishlist]
      summary: Get customer wishlist
      responses:
        "200":
          description: Wishlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "201":
          description: Added to wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
  /wishlist/{id}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        niche:
          type: string
        imageUrl:
          type: ["string", "null"]
        inStock:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - price
        - niche
        - inStock
        - createdAt
    ProductInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        niche:
          type: string
        imageUrl:
          type: string
        inStock:
          type: boolean
      required:
        - title
        - price
        - niche
    ProductUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        niche:
          type: string
        imageUrl:
          type: string
        inStock:
          type: boolean
...
    WishlistItem:
      type: object
      properties:
        id:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - product
        - createdAt
    WishlistInput:
      type: object
      properties:
        productId:
          type: integer
      required:
        - productId
...