# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Social Seller Toolkit API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  niche: string;
  /** @nullable */
  imageUrl?: string | null;
  inStock: boolean;
  createdAt: Date;
}