# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Social Seller Toolkit API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  title: string;
  description?: string;
  price: number;
  niche: string;
  imageUrl?: string;
  inStock?: boolean;
}